DROTH-3667 remove location of filter object
